First Component Validation
testing a validation function
