+++ resources/js/Components/FormCity.jsx
@@ -91,166 +91,166 @@ function classNames(...classes) {
     return classes.filter(Boolean).join(' ')
 }
 
-export function getFormData(){
-    
+export function getFormData() {
+
     var city = document.getElementById('city').value;
     var telephone = document.getElementById('phone').value;
     var fullname = document.getElementById('fullname').value;
     var province = dropdown_selected;
     var email = document.getElementById('email').value;
     var obj = {
-        "city" : city,
-        "telephone" : telephone,
-        "province" : province.name,
+        "city": city,
+        "telephone": telephone,
+        "province": province.name,
         "fullName": fullname,
         "email": email,
     }
     return obj;
 }
-function setData(x){
+function setData(x) {
     dropdown_selected = x;
 }
 
 function FormCity({ auth, laravelVersion, phpVersion }) {
     const [selected, setSelected] = useState(people[9])
 
-    const handleOnChange = (selected) =>{
+    const handleOnChange = (selected) => {
         setSelected(selected);
         setData(selected);
         return true;
     }
     return (
         <>
-        
-        <Listbox Label={"province"} value={selected} onChange={handleOnChange}>
-            {({ open }) => (
-                <>
 
-                    <Listbox.Label className="block text-sm font-medium leading-6 text-gray-900">Province</Listbox.Label>
-                    <div className="relative mt-2">
-                        <Listbox.Button className="relative w-full cursor-default rounded-md bg-white py-1.5 pl-3 pr-10 text-left text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:outline-none focus:ring-2 focus:ring-indigo-500 sm:text-sm sm:leading-6">
-                            <span className="flex items-center">
-                                <img src={selected.avatar} alt="" className="h-5 w-5 flex-shrink-0 rounded-full" />
-                                <span className="ml-3 block truncate">{selected.name}</span>
-                            </span>
-                            <span className="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2">
-                                <ChevronUpDownIcon className="h-5 w-5 text-gray-400" aria-hidden="true" />
-                            </span>
-                        </Listbox.Button>
+            <Listbox Label={"province"} value={selected} onChange={handleOnChange}>
+                {({ open }) => (
+                    <>
 
-                        <Transition
-                            show={open}
-                            as={Fragment}
-                            leave="transition ease-in duration-100"
-                            leaveFrom="opacity-100"
-                            leaveTo="opacity-0"
-                        >
-                            <Listbox.Options className="absolute z-10 mt-1 max-h-56 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                                {people.map((person) => (
-                                    <Listbox.Option
-                                        key={person.id}
-                                        className={({ active }) => classNames(
-                                            active ? 'bg-indigo-600 text-white' : 'text-gray-900',
-                                            'relative cursor-default select-none py-2 pl-3 pr-9'
-                                        )}
-                                        value={person}
-                                    >
-                                        {({ selected, active }) => (
-                                            <>
-                                                <div className="flex items-center">
-                                                    <img src={person.avatar} alt="" className="h-5 w-5 flex-shrink-0 rounded-full" />
-                                                    <span
-                                                        className={classNames(selected ? 'font-semibold' : 'font-normal', 'ml-3 block truncate')}
-                                                    >
-                                                        {person.name}
-                                                    </span>
-                                                </div>
+                        <Listbox.Label className="block text-sm font-medium leading-6 text-gray-900">Province</Listbox.Label>
+                        <div className="relative mt-2">
+                            <Listbox.Button className="relative w-full cursor-default rounded-md bg-white py-1.5 pl-3 pr-10 text-left text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:outline-none focus:ring-2 focus:ring-indigo-500 sm:text-sm sm:leading-6">
+                                <span className="flex items-center">
+                                    <img src={selected.avatar} alt="" className="h-5 w-5 flex-shrink-0 rounded-full" />
+                                    <span className="ml-3 block truncate">{selected.name}</span>
+                                </span>
+                                <span className="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2">
+                                    <ChevronUpDownIcon className="h-5 w-5 text-gray-400" aria-hidden="true" />
+                                </span>
+                            </Listbox.Button>
 
-                                                {selected ? (
-                                                    <span
-                                                        className={classNames(
-                                                            active ? 'text-white' : 'text-indigo-600',
-                                                            'absolute inset-y-0 right-0 flex items-center pr-4'
-                                                        )}
-                                                    >
-                                                        <CheckIcon className="h-5 w-5" aria-hidden="true" />
-                                                    </span>
-                                                ) : null}
-                                            </>
-                                        )}
-                                    </Listbox.Option>
-                                ))}
-                            </Listbox.Options>
-                        </Transition>
-                    </div>
+                            <Transition
+                                show={open}
+                                as={Fragment}
+                                leave="transition ease-in duration-100"
+                                leaveFrom="opacity-100"
+                                leaveTo="opacity-0"
+                            >
+                                <Listbox.Options className="absolute z-10 mt-1 max-h-56 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
+                                    {people.map((person) => (
+                                        <Listbox.Option
+                                            key={person.id}
+                                            className={({ active }) => classNames(
+                                                active ? 'bg-indigo-600 text-white' : 'text-gray-900',
+                                                'relative cursor-default select-none py-2 pl-3 pr-9'
+                                            )}
+                                            value={person}
+                                        >
+                                            {({ selected, active }) => (
+                                                <>
+                                                    <div className="flex items-center">
+                                                        <img src={person.avatar} alt="" className="h-5 w-5 flex-shrink-0 rounded-full" />
+                                                        <span
+                                                            className={classNames(selected ? 'font-semibold' : 'font-normal', 'ml-3 block truncate')}
+                                                        >
+                                                            {person.name}
+                                                        </span>
+                                                    </div>
 
-                    <div>
-                        <label htmlFor="City" className="block text-sm font-medium leading-6 text-gray-900">
-                            City
-                        </label>
-                        <div className="relative mt-2 rounded-md shadow-sm">
-                            <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
-                                <span className="text-gray-500 sm:text-sm">~</span>
-                            </div>
-                            <input
-                                type="text"
-                                name="City"
-                                id="city"
-                                className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                placeholder="City" />
+                                                    {selected ? (
+                                                        <span
+                                                            className={classNames(
+                                                                active ? 'text-white' : 'text-indigo-600',
+                                                                'absolute inset-y-0 right-0 flex items-center pr-4'
+                                                            )}
+                                                        >
+                                                            <CheckIcon className="h-5 w-5" aria-hidden="true" />
+                                                        </span>
+                                                    ) : null}
+                                                </>
+                                            )}
+                                        </Listbox.Option>
+                                    ))}
+                                </Listbox.Options>
+                            </Transition>
                         </div>
-                    </div>
-                    <div>
-                        <label htmlFor="phone" className="block text-sm font-medium leading-6 text-gray-900">
-                            Phone Number
-                        </label>
-                        <div className="relative mt-2 rounded-md shadow-sm">
-                            <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
-                              
+
+                        <div>
+                            <label htmlFor="City" className="block text-sm font-medium leading-6 text-gray-900">
+                                City
+                            </label>
+                            <div className="relative mt-2 rounded-md shadow-sm">
+                                <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+                                    <span className="text-gray-500 sm:text-sm">~</span>
+                                </div>
+                                <input
+                                    type="text"
+                                    name="City"
+                                    id="city"
+                                    className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                                    placeholder="City" />
                             </div>
-                            <input
-                                type="text"
-                                name="phone"
-                                id="phone"
-                                className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                placeholder="+1 (XXX) XXX-XXXX" />
                         </div>
-                    </div>
-                    <div>
-                        <label htmlFor="email" className="block text-sm font-medium leading-6 text-gray-900">
-                            Email
-                        </label>
-                        <div className="relative mt-2 rounded-md shadow-sm">
-                            <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
-                              
+                        <div>
+                            <label htmlFor="phone" className="block text-sm font-medium leading-6 text-gray-900">
+                                Phone Number
+                            </label>
+                            <div className="relative mt-2 rounded-md shadow-sm">
+                                <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+
+                                </div>
+                                <input
+                                    type="text"
+                                    name="phone"
+                                    id="phone"
+                                    className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                                    placeholder="+1 (XXX) XXX-XXXX" />
                             </div>
-                            <input
-                                type="email"
-                                name="email"
-                                id="email"
-                                className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                placeholder="johndoe@willexample.com...." />
                         </div>
-                    </div>
-                    <div>
-                        <label htmlFor="name" className="block text-sm font-medium leading-6 text-gray-900">
-                            Full Name
-                        </label>
-                        <div className="relative mt-2 rounded-md shadow-sm">
-                            <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
-                              
+                        <div>
+                            <label htmlFor="email" className="block text-sm font-medium leading-6 text-gray-900">
+                                Email
+                            </label>
+                            <div className="relative mt-2 rounded-md shadow-sm">
+                                <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+
+                                </div>
+                                <input
+                                    type="email"
+                                    name="email"
+                                    id="email"
+                                    className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                                    placeholder="johndoe@willexample.com...." />
                             </div>
-                            <input
-                                type="text"
-                                name="name"
-                                id="fullname"
-                                className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                placeholder="John Doe..." />
                         </div>
-                    </div>
-                </>
-            )}
-        </Listbox></>
+                        <div>
+                            <label htmlFor="name" className="block text-sm font-medium leading-6 text-gray-900">
+                                Full Name
+                            </label>
+                            <div className="relative mt-2 rounded-md shadow-sm">
+                                <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+
+                                </div>
+                                <input
+                                    type="text"
+                                    name="name"
+                                    id="fullname"
+                                    className="block w-full rounded-md border-0 py-1.5 pl-7 pr-20 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                                    placeholder="John Doe..." />
+                            </div>
+                        </div>
+                    </>
+                )}
+            </Listbox></>
